Vite build adds correct file paths to html
Previously, vite build would not add a relative path to the scripts in the generated html file leading to issues with Electron-builder not loading the scripts properly. By adding a base property with a value of an empty string to vite.config.js, we change vite's base file path so as to remove the base URL and forcing it to be relative to its deployment directory.